Refactor interfaces to new `nnbench.types` module
This module will be advertised as the one containing (most of) the
abstractions that the user can/must implement to customize nnbench.
